// MultiProgressControl/MultiProgressControl/generated/ManifestTypes.d.ts
/*
*This is auto generated from the ControlManifest.Input.xml file
*/

// Define IInputs and IOutputs Type. They should match with ControlManifest.
export interface IInputs {
    field1: ComponentFramework.PropertyTypes.WholeNumberProperty;
    field2: ComponentFramework.PropertyTypes.WholeNumberProperty;
    field3: ComponentFramework.PropertyTypes.WholeNumberProperty;
    field4: ComponentFramework.PropertyTypes.WholeNumberProperty;
    field5: ComponentFramework.PropertyTypes.WholeNumberProperty;
    field6: ComponentFramework.PropertyTypes.WholeNumberProperty;
    label1: ComponentFramework.PropertyTypes.StringProperty;
    label2: ComponentFramework.PropertyTypes.StringProperty;
    label3: ComponentFramework.PropertyTypes.StringProperty;
    label4: ComponentFramework.PropertyTypes.StringProperty;
    label5: ComponentFramework.PropertyTypes.StringProperty;
    label6: ComponentFramework.PropertyTypes.StringProperty;
    maxValue: ComponentFramework.PropertyTypes.WholeNumberProperty;
    headerLabel: ComponentFramework.PropertyTypes.StringProperty;
}
export interface IOutputs {
    field1?: number;
    field2?: number;
    field3?: number;
    field4?: number;
    field5?: number;
    field6?: number;
}